Move each class to its own file, apply proper formatting - for rate movies
