+++ features/mpmoviedetails/src/main/java/com/jpp/mpmoviedetails/rates/RateMovieParam.kt
@@ -1,0 +1,26 @@
+package com.jpp.mpmoviedetails.rates
+
+import android.os.Bundle
+import com.jpp.mp.common.extensions.getStringOrFail
+
+/**
+ * The initialization parameter used for
+ * [RateMovieViewModel].
+ */
+data class RateMovieParam(
+    val movieId: Double,
+    val screenTitle: String,
+    val movieImageUrl: String
+) {
+    companion object {
+        fun fromArguments(
+            arguments: Bundle?
+        ): RateMovieParam {
+            return RateMovieParam(
+                arguments.getStringOrFail("movieId").toDouble(),
+                arguments.getStringOrFail("movieTitle"),
+                arguments.getStringOrFail("movieImageUrl")
+            )
+        }
+    }
+}
